// src/app/Models/user-login.ts
export class UserLogin {
    Id: number;
    NickName: string;
    DocumentType: string;
    Document: string;
    Name: string;
    Email: string;
    NumberPhone: string;
    City: string;
    Departament: string;
    Address: string;
    Rol: number;
    Token: string;
    CountCompany: number;
}